Enhance Schema Designer: add report retrieval, update CSS positioning, and improve button accessibility
