+++ .vscode/launch.json
@@ -46,7 +46,7 @@
 				// Uncomment this to use a specified version of STS, see
 				// https://github.com/microsoft/vscode-mssql/blob/main/DEVELOPMENT.md#using-mssql_sqltoolsservice-environment-variable
 				// for more details
-				// "MSSQL_SQLTOOLSSERVICE": "<Path to STS>"
+				"MSSQL_SQLTOOLSSERVICE": "/Users/aasimkhan/src/sqltoolsservice/src/Microsoft.SqlTools.ServiceLayer/bin/Debug/net8.0/osx-arm64/publish"
 			}
 		},
 		{
