+++ src/reactviews/pages/SchemaDesigner/index.tsx
@@ -7,7 +7,7 @@ import ReactDOM from "react-dom/client";
 import "../../index.css";
 import { VscodeWebviewProvider } from "../../common/vscodeWebviewProvider";
 import { SchemaDesignerStateProvider } from "./schemaDesignerStateProvider";
-import { SchemaDesignerPage } from "./schemaDesignerContainer";
+import { SchemaDesignerPage } from "./schemaDesignerPage";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
     <VscodeWebviewProvider>
