+++ src/reactviews/pages/SchemaDesigner/schemaDesigner.css
@@ -22,6 +22,9 @@
     width: 100px;
     height: 100px;
     overflow: auto;
+    position: absolute;
+    bottom: 10px;
+    right: 10px;
 }
 
 .sd-toolbar-belt {
