+++ src/reactviews/pages/SchemaDesigner/schemaDesignerUtils.ts
@@ -7,7 +7,11 @@ import { SchemaDesignerConfig } from "azdataGraph/dist/src/ts/schemaDesigner/sch
 import * as schemaDesignerIcons from "./schemaDesignerIcons";
 const connectorIcon = require("./icons/connector.svg");
 import * as azdataGraph from "azdataGraph";
+import { SchemaDesigner } from "../../../sharedInterfaces/schemaDesigner";
 
+/**
+ * Get the schema designer colors from the current theme
+ */
 export function getSchemaDesignerColors(): azdataGraph.SchemaDesignerColors {
     const body = document.body;
     const computedStyle = getComputedStyle(body);
@@ -45,6 +49,9 @@ export function getSchemaDesignerColors(): azdataGraph.SchemaDesignerColors {
     };
 }
 
+/**
+ * Schema designer configuration
+ */
 export const config: SchemaDesignerConfig = {
     icons: {
         addTableIcon: schemaDesignerIcons.addTableIcon,
@@ -80,11 +87,77 @@ export const config: SchemaDesignerConfig = {
         _x: number,
         _y: number,
         _scale: number,
-    ): void {
-        throw new Error("Function not implemented.");
-    },
+    ): void {},
     publish(_schema: azdataGraph.Schema): void {
         throw new Error("Function not implemented.");
     },
     showToolbar: false,
 };
+
+/**
+ * Generate a new name for the column based on the existing columns
+ * @param existingColumns The existing columns in the table
+ * @returns The new column name
+ */
+export function getNextColumnName(
+    existingColumns: SchemaDesigner.Column[],
+): string {
+    let index = 1;
+    let columnName = `column_${index}`;
+    while (existingColumns.some((column) => column.name === columnName)) {
+        index++;
+        columnName = `column_${index}`;
+    }
+    return columnName;
+}
+
+/**
+ * Generate a new name for the foreign key based on the existing foreign keys
+ * @param existingEdges The existing foreign keys in the table
+ * @returns The new foreign key name
+ */
+export function getNextForeignKeyName(
+    existingEdges: SchemaDesigner.ForeignKey[],
+): string {
+    let index = 1;
+    let foreignKeyName = `FK_${index}`;
+    while (existingEdges.some((edge) => edge.name === foreignKeyName)) {
+        index++;
+        foreignKeyName = `FK_${index}`;
+    }
+    return foreignKeyName;
+}
+
+/**
+ * Get all tables in the schema except the current one
+ * @param schema
+ * @param currentTable
+ * @returns
+ */
+export function getAllTables(
+    schema: SchemaDesigner.Schema,
+    currentTable: SchemaDesigner.Table,
+): SchemaDesigner.Table[] {
+    return schema.tables
+        .filter(
+            (entity) =>
+                entity.schema !== currentTable.schema ||
+                entity.name !== currentTable.name,
+        )
+        .sort();
+}
+
+/**
+ * Get display name for the table
+ * @param schema schema
+ * @param displayName display name
+ * @returns table
+ */
+export function getTableFromDisplayName(
+    schema: SchemaDesigner.Schema,
+    displayName: string,
+): SchemaDesigner.Table {
+    return schema.tables.find(
+        (entity) => `${entity.schema}.${entity.name}` === displayName,
+    )!;
+}
